lsp: use luajit
fish: decent way to patch

doc: new keymap

chore: upd

+++ .config/fish/functions/upd-nvim.fish
@@ -1,7 +1,14 @@
 function upd-nvim
+  argparse -n upd-nvim --max-args=1 p/pull -- $argv
   if not test -d ~/b/neovim
-    mkdir ~/b/neovim -p
-    git clone https://github.com/neovim/neovim
+    git clone https://github.com/phanen/neovim ~/b/neovim
+    git remote add upstream http://github.com/neovim/neovim
+    git branch r upstream/master
+  end
+  if set -q _flag_pull
+    git pull upstream master:r
+    git rebase r master
+    return
   end
   rm -rf ~/b/neovim/build
   CC=clang CXX=clang++ cmake \
